erreurs typage

+++ src/main/java/com/atoudeft/banque/OperationDepot.java
@@ -1,9 +1,9 @@
 package com.atoudeft.banque;
 
 public class OperationDepot extends Operation {
-    private int montant = 0;
+    private double montant = 0;
 
-    public OperationDepot(int montant) {
+    public OperationDepot(double montant) {
         super.type = TypeOperation.DEPOT;
         this.montant = montant;
     }
